+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.3
+
+* feat(helpers): Resolve `thenOpen()` relative paths to `baseUrl` config.
+* feat(helpers): Make `url()` leading slash optional.
+
 # 0.1.2
 
 * Upgrade `apidox` and dependents.